Move nodes to adjustments module
